create DiskCache service
